feat(app-shell): redirect to project create without projects
