+++ packages/application-shell/src/components/application-shell/application-shell.js
@@ -12,7 +12,6 @@ import {
   SentryUserTracker,
 } from '@commercetools-frontend/sentry';
 import { ApplicationContextProvider } from '@commercetools-frontend/application-shell-connectors';
-import { encode } from 'qss';
 import { NotificationsList } from '@commercetools-frontend/react-notifications';
 import { AsyncLocaleData } from '@commercetools-frontend/i18n';
 import { i18n } from '@commercetools-frontend/ui-kit';
@@ -35,6 +34,7 @@ import GtmUserTracker from '../gtm-user-tracker';
 import NavBar, { LoadingNavBar } from '../navbar';
 import ApplicationLoader from '../application-loader';
 import ErrorApologizer from '../error-apologizer';
+import Redirector from '../redirector';
 import {
   selectProjectKeyFromLocalStorage,
   selectProjectKeyFromUrl,
@@ -349,44 +349,6 @@ RestrictedApplication.propTypes = {
   DEV_ONLY__loadAppbarMenuConfig: PropTypes.func,
   DEV_ONLY__loadNavbarMenuConfig: PropTypes.func,
 };
-
-class Redirector extends React.PureComponent {
-  static displayName = 'Redirector';
-  static propTypes = {
-    to: PropTypes.string.isRequired,
-    location: PropTypes.shape({
-      query: PropTypes.shape({
-        reason: PropTypes.oneOf(Object.values(LOGOUT_REASONS)).isRequired,
-      }),
-    }).isRequired,
-    environment: PropTypes.shape({
-      servedByProxy: PropTypes.bool.isRequired,
-    }).isRequired,
-    queryParams: PropTypes.shape({
-      reason: PropTypes.oneOf(Object.values(LOGOUT_REASONS)).isRequired,
-      redirectTo: PropTypes.string,
-    }).isRequired,
-  };
-  static defaultProps = {
-    location: {},
-  };
-  redirectTo = targetUrl => window.location.replace(targetUrl);
-  componentDidMount() {
-    // For now the authentication service runs on the same domain as the application,
-    // even on development (using the webpack dev server).
-    const authUrl = window.location.origin;
-
-    const searchQuery = {
-      ...this.props.queryParams,
-      ...(this.props.location.query || {}),
-    };
-
-    this.redirectTo(`${authUrl}/${this.props.to}?${encode(searchQuery)}`);
-  }
-  render() {
-    return null;
-  }
-}
 
 const isBooleanAsString = environmentValue =>
   environmentValue === 'true' || environmentValue === 'false';
