+++ packages/application-shell/src/components/redirector/index.js
@@ -1,0 +1,1 @@
+export { default } from './redirector';
